+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## UNRELEASED
+
+BREAKING CHANGES:
+
+- The following values have been removed from the `subnets.delegation` parameter: `Microsoft.App/testClients`, `Microsoft.Codespaces/plans` and `Microsoft.Orbital/orbitalGateways`.
+
+ENHANCEMENTS:
+
+- `subnets.service_endpoints`: updated to use the `service_endpoint` block from `hashicorp/azurerm` provider v5.0.
+- New options have been added to the `subnets.service_endpoints` parameter: `Microsoft.CognitiveServices`.
+- New options have been added to the `subnets.delegation` parameter: `Microsoft.MessagingConnectors/connectors` and `PureStorage.Block/storagePools`.
+- dependencies: updating to `v5.0.0` minimum of `hashicorp/azurerm` provider.
+
 ## 5.0.0 (June 16, 2025)
 
 BREAKING CHANGES: